backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
supabase==2.10.0
PyJWT==2.10.1
google-generativeai==0.8.3
redis==5.2.1
python-dotenv==1.0.1
httpx>=0.27,<0.28
pytest==8.3.4
pytest-asyncio==0.24.0
